add xor utility
xor entire files together, like [Monolith] but without being super
complex for no reason (seriously, Monolith is over a thousand lines of
C++ just for the file to xor the contents of two files, and the whole
thing is nearly 48k lines of code)

note this is moderately slower than Monolith, as that reads the entire
contents of both files and the output into ram, while we just read as
needed. this also differs in that the output length will be the smaller
of the two files, whereas Monolith will choose one file for the length,
and wrap around the other as needed, possibly introducing unwanted
patterns.

[Monolith]: https://monolith.sourceforge.net/